Replaced micromath with constgebra

+++ src/animator/animator.rs
@@ -1,11 +1,11 @@
 use wasm4::tracef;
 
 pub enum Transformation {
-    Rotation { start: f32, end: f32 },
+    Rotation { start: f64, end: f64 },
 }
 
 pub trait Animate {
-    fn set_rotation(&mut self, rotation: f32);
+    fn set_rotation(&mut self, rotation: f64);
     // fn animate<F: FnMut(&mut Self)>(&mut self, mut f: F) {
     //     (f)(self);
     // }
@@ -33,7 +33,7 @@ impl Animator {
         }
 
         self.elapsed += delta_time;
-        let progress = (self.elapsed as f32 / self.duration as f32).min(1.0);
+        let progress = (self.elapsed as f64 / self.duration as f64).min(1.0);
 
         match self.transformation {
             Transformation::Rotation { start, end } => {
@@ -53,6 +53,6 @@ impl Animator {
         self.is_complete
     }
 }
-fn lerp(start: f32, end: f32, t: f32) -> f32 {
+fn lerp(start: f64, end: f64, t: f64) -> f64 {
     start + (end - start) * t
 }
